move smtp functions to separate file
