+++ log_functions.py
@@ -2,9 +2,6 @@ import logging
 import pathlib
 import sys
 import typing
-import smtplib
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
 
 from dynaconf import Dynaconf
 
@@ -54,28 +51,3 @@ def log_verbosity_to_logging(verbosity: str):
 def handle_error_and_exit(error: typing.Union[str, Exception]):
     logging.error(str(error))
     sys.exit(1)
-
-
-def send_email(error: str, date_time: str = None) -> None:
-    smtp_conn = smtplib.SMTP(config.notification.smtp_server)
-    smtp_conn.login(config.notification.smtp_username, config.notification.smtp_password)
-
-    msg = MIMEMultipart()
-    msg["Subject"] = config.notification.email_subject
-    msg["From"] = config.notification.email_sender
-    msg["To"] = ", ".join(config.notification.email_recipients)
-    msg.preamble = error
-
-    if date_time:
-        error = f"{date_time}: {error}"
-
-    msg.attach(MIMEText(error, "plain"))
-
-    smtp_conn.sendmail(config.notification.smtp_username, config.notification.email_recipients, msg.as_string())
-    smtp_conn.quit()
-
-
-def can_send_email(config: Dynaconf) -> bool:
-    if "notification" in config:
-        return True
-    return False
